requirements.txt
appdirs==1.4.3
attrs==18.2.0
Babel==2.6.0
black==18.9b0
Click==7.0
coverage==4.5.2
Django==2.1.4
django-crispy-forms==1.7.2
django-phonenumber-field==2.1.0
flake8==3.6.0
homevisit==0.0.1
mccabe==0.6.1
mypy==0.650
mypy-extensions==0.4.1
phonenumbers==8.10.2
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.7
toml==0.10.0
typed-ast==1.1.0
